+++ CHANGELOG.md
@@ -7,6 +7,7 @@ For detailed information check [Aerofly Weather's releases on Github](https://gi
 * :gift: Add CLI parameters `--time=<X>` and `--date=<X>`
 * :gift: Add General Data Protection Regulation (GDPR) compliance statement
 * :gift: Add examples for desktop integration
+* :gift: Add `aerofly-make-batch` to add desktop shortcut for `aerofly-metar`
 * :wrench: Improve internal documentation and structure
 * :bomb: Changed internal METAR object structure to match https://api.checkwx.com/#31-single